Move tag-store to an own file.
